rename and move notification cache to its own file
